requirements.txt
Flask
flask-cors
requests
python-dotenv
openai
yt-dlp
indic-transliteration
